+++ hydra-node/src/Hydra/Node/Run.hs
@@ -1,20 +1,31 @@
+{-# LANGUAGE DisambiguateRecordFields #-}
 {-# LANGUAGE DuplicateRecordFields #-}
 
 module Hydra.Node.Run where
 
 import Hydra.Prelude hiding (fromList)
 
+import Cardano.Ledger.BaseTypes qualified as Ledger
+import Cardano.Ledger.Crypto qualified as Ledger
+import Cardano.Ledger.Shelley.API qualified as Shelley
 import Hydra.API.Server (Server (..), withAPIServer)
 import Hydra.API.ServerOutput (ServerOutput (..))
 import Hydra.Cardano.Api (
+  GenesisParameters (..),
   ProtocolParametersConversionError,
   ShelleyBasedEra (..),
+  StandardCrypto,
+  SystemStart (..),
   toLedgerPParams,
  )
+import Hydra.Cardano.Api qualified as Shelley
 import Hydra.Chain (maximumNumberOfParties)
 import Hydra.Chain.CardanoClient (QueryPoint (..), queryGenesisParameters)
 import Hydra.Chain.Direct (loadChainContext, mkTinyWallet, withDirectChain)
+import Hydra.Chain.Direct.Fixture (defaultGlobals)
 import Hydra.Chain.Direct.State (initialChainState)
+import Hydra.Chain.Offline (withOfflineChain)
+import Hydra.HeadId (HeadId (..))
 import Hydra.HeadLogic (
   Environment (..),
   Event (..),
@@ -30,6 +41,7 @@ import Hydra.Ledger.Cardano.Configuration (
 import Hydra.Logging (Verbosity (..), traceWith, withTracer)
 import Hydra.Logging.Messages (HydraLog (..))
 import Hydra.Logging.Monitoring (withMonitoring)
+import Hydra.Network (NodeId (NodeId))
 import Hydra.Network.Authenticate (Authenticated (Authenticated))
 import Hydra.Network.Message (Connectivity (..))
 import Hydra.Node (
@@ -37,6 +49,7 @@ import Hydra.Node (
   checkHeadState,
   createNodeState,
   initEnvironment,
+  initEnvironmentOffline,
   loadState,
   runHydraNode,
  )
@@ -46,9 +59,13 @@ import Hydra.Options (
   ChainConfig (..),
   InvalidOptions (..),
   LedgerConfig (..),
+  OfflineConfig (..),
+  RunOfflineOptions (..),
   RunOptions (..),
   validateRunOptions,
  )
+import Hydra.Options.Offline qualified as OfflineOptions
+import Hydra.Options.Online qualified as OnlineOptions
 import Hydra.Persistence (createPersistenceIncremental)
 
 data ConfigurationException
@@ -66,6 +83,67 @@ explain = \case
   ConfigurationException err ->
     "Incorrect protocol parameters configuration provided: " <> show err
 
+runOffline :: RunOfflineOptions -> IO ()
+runOffline opts = do
+  either (throwIO . InvalidOptionException) pure $ OfflineOptions.validateRunOfflineOptions opts
+  let RunOfflineOptions{verbosity, monitoringPort, persistenceDir, offlineConfig} = opts
+  env@Environment{party, otherParties, signingKey, contestationPeriod} <- initEnvironmentOffline opts
+
+  withTracer verbosity $ \tracer' ->
+    withMonitoring monitoringPort tracer' $ \tracer -> do
+      traceWith tracer (NodeOfflineOptions opts)
+      eq@EventQueue{putEvent} <- createEventQueue
+      let RunOfflineOptions{ledgerConfig} = opts
+      protocolParams <- readJsonFileThrow protocolParametersFromJson (cardanoLedgerProtocolParametersFile ledgerConfig)
+      pparams <- case toLedgerPParams ShelleyBasedEraBabbage protocolParams of
+        Left err -> throwIO (ConfigurationException err)
+        Right bpparams -> pure bpparams
+
+      globals <- loadGlobalsFromGenesis (ledgerGenesisFile offlineConfig)
+
+      withCardanoLedger pparams globals $ \ledger -> do
+        persistence <- createPersistenceIncremental $ persistenceDir <> "/state"
+        (hs, chainStateHistory) <- loadState (contramap Node tracer) persistence initialChainState
+
+        checkHeadState (contramap Node tracer) env hs
+        nodeState <- createNodeState hs
+        -- Chain
+        let withChain cont =
+              let headId = UnsafeHeadId "HeadId"
+               in withOfflineChain (contramap DirectChain tracer) offlineConfig globals headId party contestationPeriod chainStateHistory (putEvent . OnChainEvent) cont
+        withChain $ \chain -> do
+          -- API
+          let RunOfflineOptions{host, port} = opts
+              peers = []
+              nodeId = NodeId "offline"
+              putNetworkEvent (Authenticated msg otherParty) = putEvent $ NetworkEvent defaultTTL otherParty msg
+              RunOfflineOptions{apiHost, apiPort} = opts
+          apiPersistence <- createPersistenceIncremental $ persistenceDir <> "/server-output"
+          withAPIServer apiHost apiPort party apiPersistence (contramap APIServer tracer) chain pparams (putEvent . ClientEvent) $ \server -> do
+            -- Network
+            let networkConfiguration = NetworkConfiguration{persistenceDir, signingKey, otherParties, host, port, peers, nodeId}
+            withNetwork tracer (connectionMessages server) networkConfiguration putNetworkEvent $ \hn -> do
+              -- Main loop
+              runHydraNode (contramap Node tracer) $
+                HydraNode
+                  { eq
+                  , hn
+                  , nodeState
+                  , oc = chain
+                  , server
+                  , ledger
+                  , env
+                  , persistence
+                  }
+ where
+  connectionMessages Server{sendOutput} = \case
+    Connected nodeid -> sendOutput $ PeerConnected nodeid
+    Disconnected nodeid -> sendOutput $ PeerDisconnected nodeid
+
+  withCardanoLedger protocolParams globals action =
+    let ledgerEnv = newLedgerEnv protocolParams
+     in action (Ledger.cardanoLedger globals ledgerEnv)
+
 run :: RunOptions -> IO ()
 run opts = do
   either (throwIO . InvalidOptionException) pure $ validateRunOptions opts
@@ -80,15 +158,23 @@ run opts = do
       pparams <- case toLedgerPParams ShelleyBasedEraBabbage protocolParams of
         Left err -> throwIO (ConfigurationException err)
         Right bpparams -> pure bpparams
-      withCardanoLedger chainConfig pparams $ \ledger -> do
+
+      let DirectChainConfig{networkId, nodeSocket} = chainConfig
+
+      globals <- newGlobals =<< queryGenesisParameters networkId nodeSocket QueryTip
+
+      withCardanoLedger pparams globals $ \ledger -> do
         persistence <- createPersistenceIncremental $ persistenceDir <> "/state"
         (hs, chainStateHistory) <- loadState (contramap Node tracer) persistence initialChainState
+
         checkHeadState (contramap Node tracer) env hs
         nodeState <- createNodeState hs
         -- Chain
-        ctx <- loadChainContext chainConfig party hydraScriptsTxId
-        wallet <- mkTinyWallet (contramap DirectChain tracer) chainConfig
-        withDirectChain (contramap DirectChain tracer) chainConfig ctx wallet chainStateHistory (putEvent . OnChainEvent) $ \chain -> do
+        let withChain cont = do
+              ctx <- loadChainContext chainConfig party hydraScriptsTxId
+              wallet <- mkTinyWallet (contramap DirectChain tracer) chainConfig
+              withDirectChain (contramap DirectChain tracer) chainConfig ctx wallet chainStateHistory (putEvent . OnChainEvent) cont
+        withChain $ \chain -> do
           -- API
           let RunOptions{host, port, peers, nodeId} = opts
               putNetworkEvent (Authenticated msg otherParty) = putEvent $ NetworkEvent defaultTTL otherParty msg
@@ -115,12 +201,63 @@ run opts = do
     Connected nodeid -> sendOutput $ PeerConnected nodeid
     Disconnected nodeid -> sendOutput $ PeerDisconnected nodeid
 
-  withCardanoLedger chainConfig protocolParams action = do
-    let DirectChainConfig{networkId, nodeSocket} = chainConfig
-    globals <- newGlobals =<< queryGenesisParameters networkId nodeSocket QueryTip
+  withCardanoLedger protocolParams globals action =
     let ledgerEnv = newLedgerEnv protocolParams
-    action (Ledger.cardanoLedger globals ledgerEnv)
+     in action (Ledger.cardanoLedger globals ledgerEnv)
 
 identifyNode :: RunOptions -> RunOptions
-identifyNode opt@RunOptions{verbosity = Verbose "HydraNode", nodeId} = opt{verbosity = Verbose $ "HydraNode-" <> show nodeId}
+identifyNode opt@RunOptions{verbosity = Verbose "HydraNode", nodeId} = opt{OnlineOptions.verbosity = Verbose $ "HydraNode-" <> show nodeId}
 identifyNode opt = opt
+
+loadGlobalsFromGenesis :: Maybe FilePath -> IO Shelley.Globals
+loadGlobalsFromGenesis ledgerGenesisFile = do
+  shelleyGenesis <- case ledgerGenesisFile of
+    Nothing -> pure Nothing
+    Just filePath -> Just <$> readJsonFileThrow (parseJSON @(Ledger.ShelleyGenesis StandardCrypto)) filePath
+  systemStart <- maybe (SystemStart <$> getCurrentTime) (pure . SystemStart . Ledger.sgSystemStart) shelleyGenesis
+
+  let genesisParameters = fromShelleyGenesis <$> shelleyGenesis
+
+  globals <-
+    maybe
+      (pure $ defaultGlobals{Ledger.systemStart = systemStart})
+      newGlobals
+      genesisParameters
+
+  pure globals
+
+-- | Taken from Cardano.Api.GenesisParameters, a private module in cardano-api
+fromShelleyGenesis :: Shelley.ShelleyGenesis Ledger.StandardCrypto -> GenesisParameters Shelley.ShelleyEra
+fromShelleyGenesis
+  sg@Shelley.ShelleyGenesis
+    { Shelley.sgSystemStart
+    , Shelley.sgNetworkMagic
+    , Shelley.sgActiveSlotsCoeff
+    , Shelley.sgSecurityParam
+    , Shelley.sgEpochLength
+    , Shelley.sgSlotsPerKESPeriod
+    , Shelley.sgMaxKESEvolutions
+    , Shelley.sgSlotLength
+    , Shelley.sgUpdateQuorum
+    , Shelley.sgMaxLovelaceSupply
+    , Shelley.sgGenDelegs = _ -- unused, might be of interest
+    , Shelley.sgInitialFunds = _ -- unused, not retained by the node
+    , Shelley.sgStaking = _ -- unused, not retained by the node
+    } =
+    GenesisParameters
+      { protocolParamSystemStart = sgSystemStart
+      , protocolParamNetworkId = Shelley.fromNetworkMagic $ Shelley.NetworkMagic sgNetworkMagic
+      , protocolParamActiveSlotsCoefficient =
+          Ledger.unboundRational
+            sgActiveSlotsCoeff
+      , protocolParamSecurity = fromIntegral sgSecurityParam
+      , protocolParamEpochLength = sgEpochLength
+      , protocolParamSlotLength = Shelley.fromNominalDiffTimeMicro sgSlotLength
+      , protocolParamSlotsPerKESPeriod = fromIntegral sgSlotsPerKESPeriod
+      , protocolParamMaxKESEvolutions = fromIntegral sgMaxKESEvolutions
+      , protocolParamUpdateQuorum = fromIntegral sgUpdateQuorum
+      , protocolParamMaxLovelaceSupply =
+          Shelley.Lovelace
+            (fromIntegral sgMaxLovelaceSupply)
+      , protocolInitialUpdateableProtocolParameters = Shelley.sgProtocolParams sg
+      }